Paste code, getting impatient

+++ migrate.ps1
@@ -103,7 +103,7 @@ Get-ChildItem -Path "content" -Filter "*.md" -Recurse | ForEach-Object {
   # Convert index files
   if ($file.BaseName -eq $file.Directory.BaseName) {
     
-    # Rename-Item -Path $file.FullName -NewName "index.md"
+    Rename-Item -Path $file.FullName -NewName "index.md"
 
     # Define a string that matches the old directory except \content\ is replaced with \assets\
     $assetDir = $file.DirectoryName -replace "content", "assets"
@@ -112,6 +112,22 @@ Get-ChildItem -Path "content" -Filter "*.md" -Recurse | ForEach-Object {
         # Move all files from the asset directory to the new directory
         Move-Item -Path "$assetDir\*" -Destination $file.DirectoryName
         # Delete the old asset directory
+        Remove-Item -Path $assetDir
+    }
+  }
+}
+
+(Get-ChildItem -Path "content\generative-works" -Filter "*.md" -Recurse) | ForEach-Object { 
+  $file = $_
+  $content = Get-Content -Path $file.FullName
+  $content = $content -replace "picture: .*\/", "picture: "
+  $content = $content -replace "thumbnail: .*\/", "thumbnail: "
+  Set-Content -Path $file.FullName -Value $content
+  if ($file.BaseName -eq $file.Directory.BaseName) {    
+    Rename-Item -Path $file.FullName -NewName "index.md"
+    $assetDir = $file.DirectoryName -replace "content", "assets"
+    if (Test-Path -Path $assetDir) {
+        Move-Item -Path "$assetDir\*" -Destination $file.DirectoryName
         Remove-Item -Path $assetDir
     }
   }
